+++ pegincontract/contract_test.go
@@ -2,22 +2,7 @@ package pegincontract
 
 import (
 	"encoding/hex"
-	"encoding/json"
-	"fmt"
-	"os/exec"
-	"strings"
 	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
-
-	"github.com/btcsuite/btcd/btcec"
-	"github.com/btcsuite/btcd/chaincfg"
-	"github.com/btcsuite/btcd/txscript"
-	"github.com/vulpemventures/go-elements/elementsutil"
-	"github.com/vulpemventures/go-elements/network"
-	"github.com/vulpemventures/go-elements/payment"
-	"github.com/vulpemventures/go-elements/pegin"
 )
 
 type NetworkType int
@@ -135,174 +120,4 @@ func TestCalculateContract(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestClaimPegin(t *testing.T) {
-	t.Skip("skipping test since it requires nigiri")
-
-	isDynaFedEnabled := false
-	federationScript := "51"
-	fedpegScriptBytes, err := hex.DecodeString(federationScript)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	privateKey, err := btcec.NewPrivateKey(btcec.S256())
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	liquidNetwork, btcNetwork := getNetworkParams(RegtestNet)
-
-	claimScript, err := pegin.ClaimWitnessScript(
-		privateKey.PubKey().SerializeCompressed(),
-		liquidNetwork,
-	)
-	t.Log(fmt.Sprintf("claimScript: %v", hex.EncodeToString(claimScript)))
-
-	contract, err := Calculate(
-		fedpegScriptBytes,
-		claimScript,
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-	t.Log(fmt.Sprintf("contract: %v", hex.EncodeToString(contract)))
-
-	mainChainAddress, err := pegin.MainChainAddress(
-		contract,
-		btcNetwork,
-		isDynaFedEnabled,
-		fedpegScriptBytes,
-	)
-	t.Log(fmt.Sprintf("mainchainAddress: %v", mainChainAddress))
-
-	output := outputCommand("nigiri", "faucet", mainChainAddress)
-	btcTxID := strings.TrimPrefix(strings.TrimSpace(string(output[:])), "txId: ")
-	t.Log(fmt.Sprintf("btcTxID: %v", btcTxID))
-
-	time.Sleep(3 * time.Second)
-
-	jsonOut := outputCommand("nigiri", "rpc", "gettransaction", btcTxID)
-	btcTxHex := getValueByKey(jsonOut, "hex")
-	t.Log(fmt.Sprintf("btcTxHex: %v", btcTxHex))
-	btcTxBytes, err := hex.DecodeString(btcTxHex)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	arg := fmt.Sprintf("[\"%v\"]", btcTxID)
-	btcTxOutProof := outputCommand("nigiri", "rpc", "gettxoutproof", arg)
-	t.Log(fmt.Sprintf("btcTxOutProof: %v", strings.TrimSpace(string(btcTxOutProof[:]))))
-	btcTxOutProofBytes, err := hex.DecodeString(strings.TrimSpace(string(btcTxOutProof[:])))
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	peggedAssetBytes, err := hex.DecodeString(network.Regtest.AssetID)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	var lbtc = append(
-		[]byte{0x01},
-		elementsutil.ReverseBytes(peggedAssetBytes)...,
-	)
-
-	parentBlockHash := "0f9188f13cb7b2c71f2a335e3a4fc328bf5beb436012afca590b1a11466e2206"
-	parentBlockHashBytes, err := hex.DecodeString(parentBlockHash)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	claimTx, err := pegin.Claim(
-		btcNetwork,
-		isDynaFedEnabled,
-		lbtc,
-		parentBlockHashBytes,
-		fedpegScriptBytes,
-		contract,
-		btcTxBytes,
-		btcTxOutProofBytes,
-		claimScript,
-		1,
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	//SIGN
-	_, amount, err := pegin.GetPeginTxOutIndexAndAmount(
-		btcTxBytes,
-		fedpegScriptBytes,
-		contract,
-		btcNetwork,
-		isDynaFedEnabled,
-	)
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	finalValue, err := elementsutil.SatoshiToElementsValue(uint64(amount))
-	if err != nil {
-		t.Fatal(err)
-	}
-	p := payment.FromPublicKey(privateKey.PubKey(), liquidNetwork, nil)
-	sigHash := claimTx.HashForWitnessV0(
-		0,
-		p.Script,
-		finalValue,
-		txscript.SigHashAll,
-	)
-	sig, err := privateKey.Sign(sigHash[:])
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	sigWithHashType := append(sig.Serialize(), byte(txscript.SigHashAll))
-	witness := make([][]byte, 0)
-	witness = append(witness, sigWithHashType[:])
-	witness = append(witness, privateKey.PubKey().SerializeCompressed())
-	claimTx.Inputs[0].Witness = witness
-	claimHex, err := claimTx.ToHex()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	output = outputCommand("nigiri", "rpc", "--liquid", "sendrawtransaction", claimHex)
-	txID := strings.TrimSpace(string(output[:]))
-	t.Log(fmt.Sprintf("claimTxID: %v", txID))
-
-	assert.NotEmpty(t, txID)
-}
-
-func getNetworkParams(
-	networkType NetworkType,
-) (*network.Network, *chaincfg.Params) {
-	var liquidNetwork *network.Network
-	var btcNetwork *chaincfg.Params
-	switch networkType {
-	case MainNet:
-		liquidNetwork = &network.Liquid
-		btcNetwork = &chaincfg.MainNetParams
-	case RegtestNet:
-		liquidNetwork = &network.Regtest
-		btcNetwork = &chaincfg.RegressionNetParams
-	}
-	return liquidNetwork, btcNetwork
-}
-
-func outputCommand(name string, arg ...string) []byte {
-	cmd := exec.Command(name, arg...)
-	bytes, err := cmd.Output()
-	if err != nil {
-		fmt.Println(err.Error())
-	}
-	return bytes
-}
-
-func getValueByKey(JSONobject []byte, key string) string {
-	var data map[string]interface{}
-	json.Unmarshal(JSONobject, &data)
-	return data[key].(string)
 }
